Fixes to most tests.
