+++ app/models/match.rb
@@ -1,5 +1,6 @@
 # frozen_string_literal: true
 
 class Match < ApplicationRecord
-  belongs_to :users, dependent: :destroy, optional: true
+  belongs_to :users, optional: true
+  belongs_to :preference_form, optional: true
 end
