+++ app/models/preference.rb
@@ -1,4 +1,5 @@
 # frozen_string_literal: true
 
 class Preference < ApplicationRecord
+  belongs_to :preference_form
 end
