+++ app/models/preference_form.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class PreferenceForm < ApplicationRecord
-  has_many :questions, dependent: :destroy
-  has_many :preferences, dependent: :destroy
-  has_many :matches, dependent: :destroy
+  has_many :questions, dependent: :delete_all
+  has_many :preferences, dependent: :delete_all
+  has_many :matches, dependent: :delete_all
 end
