+++ spec/controllers/answers_controller_spec.rb
@@ -69,25 +69,3 @@ RSpec.describe 'User actions', type: :system do
   end
 end
   
-def director_login
-  # Sign up with new account
-  visit root_path
-  click_link 'Sign up'
-  fill_in 'user[email]', with: 'user_1@gmail.com'
-  fill_in 'Password', with: '12345'
-  click_button 'Sign up'
-
-  # Approve user
-  @user = User.last
-  @user.approved = TRUE
-  @user.save
-
-  # Make the last account that signed up a Director in order to access the matches pages.
-  @director_user = User.last
-  @director_user.remove_role :chair
-  @director_user.add_role :director
-
-  # Confirm the email
-  open_email 'user_1@gmail.com'
-  click_first_link_in_email
-end
