+++ spec/controllers/matches_controller_spec.rb
@@ -22,12 +22,7 @@ RSpec.describe 'Controller Test', type: :system do
     @pref_3 = Preference.create(id: 3, selector_id: 2, selected_id: 1, preference_form_id: 1, pref_type: 'Preference', rating: 5)
     @pref_4 = Preference.create(id: 4, selector_id: 2, selected_id: 3, preference_form_id: 1, pref_type: 'Preference', rating: 3)
     @pref_5 = Preference.create(id: 5, selector_id: 3, selected_id: 1, preference_form_id: 1, pref_type: 'Preference', rating: 3)
-
-    # Match Entries
-    @match_1 = Match.create(id: 1, user_id: 1, matched_id: nil)
-    @match_2 = Match.create(id: 2, user_id: 2, matched_id: nil)
-    @match_3 = Match.create(id: 3, user_id: 3, matched_id: nil)
-
+    
     # Mock Q/A
     @q_1 = Question.create(id:1, preference_form_id: 1, question: 'Is this a quesiton?', question_type: 'Short Answer')
     @q_2 = Question.create(id:2, preference_form_id: 1, question: 'Is this a quesiton 2?', question_type: 'Multiple Choice')
@@ -38,6 +33,13 @@ RSpec.describe 'Controller Test', type: :system do
     @a_2 = Answer.create(user_id: 1, question_id: 2, preference_form_id: 1, answer_type: 'Multiple Choice', multiple_choice: 'Possibly.')
     @a_3 = Answer.create(user_id: 1, question_id: 3, preference_form_id: 1, answer_type: 'True/False', true_false: true)
     @a_4 = Answer.create(user_id: 1, question_id: 4, preference_form_id: 1, answer_type: 'Numeric', numeric: 4)
+  end
+  
+  before(:each) do
+    # Match Entries
+    @match_1 = Match.create(id: 1, user_id: 1, matched_id: nil)
+    @match_2 = Match.create(id: 2, user_id: 2, matched_id: nil)
+    @match_3 = Match.create(id: 3, user_id: 3, matched_id: nil)
   end
 
   describe 'proper message after a match has been made' do
@@ -79,7 +81,9 @@ RSpec.describe 'Director', type: :system do
     @pref_4 = Preference.create(id: 4, selector_id: 2, selected_id: 3, preference_form_id: 1, pref_type: 'Preference', rating: 3)
     @pref_5 = Preference.create(id: 5, selector_id: 3, selected_id: 1, preference_form_id: 1, pref_type: 'Preference', rating: 3)
     @pref_6 = Preference.create(id: 6, selector_id: 4, selected_id: 5, preference_form_id: 1, pref_type: 'Preference', rating: 3)
-    
+  end
+
+    before(:each) do
     # Match Entries
     @match_1 = Match.create(id: 1, user_id: 1, matched_id: nil)
     @match_2 = Match.create(id: 2, user_id: 2, matched_id: nil)
@@ -192,28 +196,4 @@ RSpec.describe 'Director', type: :system do
   after(:all) do
     DatabaseCleaner.clean_with(:truncation)
   end
-end
-
-# Helper function to login a user and make them a Director (in order to access matches pages)
-def director_login
-  # Sign up with a new account.
-  visit root_path
-  click_link 'Sign up'
-  fill_in 'user[email]', with: 'director@gmail.com'
-  fill_in 'Password', with: '12345'
-  click_button 'Sign up'
-
-  # Approve user
-  @user = User.last
-  @user.approved = TRUE
-  @user.save
-
-  # Make the last account that signed up a Director in order to access the matches pages.
-  @director_user = User.last
-  @director_user.remove_role :chair
-  @director_user.add_role :director
-
-  # Confirm the email.
-  open_email 'director@gmail.com'
-  click_first_link_in_email
 end
