+++ spec/controllers/profiles_controller_spec.rb
@@ -89,20 +89,3 @@ RSpec.describe 'User Working on profile', type: :system do
 
 end
   
-def user_login
-  # Sign up with new account
-  visit root_path
-  click_link 'Sign up'
-  fill_in 'user[email]', with: 'user@gmail.com'
-  fill_in 'Password', with: '12345'
-  click_button 'Sign up'
-
-  #Approve user
-  @user = User.last
-  @user.approved = TRUE
-  @user.save
-
-  # Confirm the email
-  open_email 'user@gmail.com'
-  click_first_link_in_email
-end
