+++ spec/controllers/public_form_spec.rb
@@ -4,11 +4,7 @@ require 'rails_helper'
 # Acceptance Test User Requirement
 RSpec.describe 'User actions', type: :system do
     it 'create, edit, and delete preference form' do
-      visit new_profile_path
-      click_link "Sign up"
-      fill_in "user[email]", with: "g@gmail.com"
-      fill_in "Password", with: "12345"
-      click_button "Sign up"
+      user_login()
   
       visit preference_forms_path
       click_link "Create New Form"
