--- spec/models/role_spec.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-require 'rails_helper'
-
-RSpec.describe Role, type: :model do
-  pending "add some examples to (or delete) #{__FILE__}"
-end
